DB-GPT/requirements.txt
abstract_singleton==1.0.1
accelerate==1.2.1
aiofiles==24.1.0
aiohttp==3.8.4
aiolimiter==1.2.1
aiosignal==1.3.2
alembic==1.12.0
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.7.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==24.3.0
auto_gpt_plugin_template==0.0.3
babel==2.16.0
backoff==2.2.1
bce-python-sdk==0.9.23
bcrypt==4.2.1
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
bleach==6.2.0
blis==0.7.11
bs4==0.0.2
build==1.2.2.post1
CacheControl==0.14.1
cachetools==5.5.0
catalogue==2.0.10
cattrs==24.1.2
certifi==2024.12.14
cffi==1.17.1
chardet==5.1.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.23
cleo==2.1.0
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cpm-kernels==1.0.11
crashtest==0.4.1
cryptography==44.0.0
cycler==0.12.1
cymem==2.0.10
dashscope==1.20.14
dataclasses-json==0.6.7
-e git+ssh://git@github.com/Laqcce-cao/DB-GPT.git@624328a0cac91beacc9bbff503e6ebfdd9cf9333#egg=dbgpt
dbgpt-tugraph-plugins==0.1.0rc1
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.9
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
duckdb==1.1.3
duckdb-engine==0.9.1
dulwich==0.21.7
durationpy==0.9
et_xmlfile==2.0.0
executing==2.1.0
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fschat==0.2.36
fsspec==2024.10.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.37.0
googleapis-common-protos==1.66.0
graphviz==0.20.3
grpcio==1.68.1
gTTS==2.3.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
importlib-resources==5.12.0
importlib_metadata==8.5.0
iniconfig==2.0.0
installer==0.7.0
ipykernel==6.29.5
ipython==8.30.0
isoduration==20.11.0
jaraco.classes==3.4.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==24.3.1
kiwisolver==1.4.7
kubernetes==31.0.0
langchain==0.3.12
langchain-community==0.3.12
langchain-core==0.3.25
langchain-openai==0.2.12
langchain-text-splitters==0.3.3
langcodes==3.5.0
langsmith==0.2.3
language_data==1.3.0
latex2mathml==3.77.0
libro==0.1.26
libro-ai==0.1.13
libro-core==0.1.6
libro-server==0.1.10
libro-sql==0.1.11
lsprotocol==2023.0.1
lxml==5.3.0
lyric-js-worker==0.1.5
lyric-py==0.1.5
lyric-py-worker==0.1.5
lyric-task==0.1.5
Mako==1.3.8
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.2
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mmh3==5.0.1
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.17
murmurhash==1.0.11
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
neo4j==5.27.0
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.19
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.4
onnxruntime==1.20.1
openai==1.57.4
openpyxl==3.1.2
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.12
overrides==7.7.0
packaging==24.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pathlib_abc==0.1.1
pathy==0.11.0
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.9.0
pillow==11.0.0
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
posthog==3.7.4
preshed==3.0.9
prettytable==3.12.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.1
psutil==5.9.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
pygls==1.3.1
Pygments==2.18.0
PyJWT==2.8.0
Pympler==1.1
PyMySQL==1.1.1
pyparsing==3.2.0
pypdf==5.1.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.0
python-json-logger==3.2.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qianfan==0.4.12.2
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rocksdict==0.3.25
rpds-py==0.22.3
rsa==4.9
ruff==0.8.3
ruff-lsp==0.0.59
safetensors==0.4.5
schedule==1.2.2
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.3.1
sentencepiece==0.2.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
snowflake-id==1.0.2
soupsieve==2.6
spacy==3.7.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.28
sqlparse==0.4.4
srsly==2.5.0
stack-data==0.6.3
starlette==0.41.3
svgwrite==1.4.3
sympy==1.13.3
tenacity==8.3.0
termcolor==2.5.0
terminado==0.18.1
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.20.3
tomlkit==0.13.2
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
trove-classifiers==2024.10.21.16
typeguard==4.4.1
typer==0.9.4
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.34.0
virtualenv==20.28.0
wasabi==1.1.3
watchfiles==1.0.3
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.3.4
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
wrapt==1.17.0
xattr==1.1.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.18.3
zhipuai==2.1.5.20241204
zipp==3.21.0
